Renamed content_scan to scan_content
