+++ src/lib/scan/scan_content.js
